+++ src/main/java/com/ravid/clothes_marketplace/app/errors/ResponseExceptionHandler.java
@@ -1,6 +1,5 @@
 package com.ravid.clothes_marketplace.app.errors;
 
-import java.math.BigDecimal;
 import java.util.NoSuchElementException;
 
 import org.springframework.http.HttpStatus;
@@ -29,20 +28,20 @@ public class ResponseExceptionHandler extends ResponseEntityExceptionHandler{
     @ExceptionHandler(JWTVerificationException.class)
     public final ResponseEntity<ErrorResponseDTO> handleJwtErrors(JWTVerificationException ex, WebRequest request) {
         return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(new ErrorResponseDTO().error("UNAUTHORIZED")
-                        .exception(ex.getMessage()).status(BigDecimal.valueOf(401)));
+                        .exception(ex.getMessage()).status(401));
     }
 
     @ExceptionHandler(AccessDeniedException.class)
     public final ResponseEntity<ErrorResponseDTO> handleAccessDeniedException(AccessDeniedException ex, WebRequest request) {
         return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(new ErrorResponseDTO().error("UNAUTHORIZED")
-                        .exception(ex.getMessage()).status(BigDecimal.valueOf(401)));
+                        .exception(ex.getMessage()).status(401));
     }
 
     @ExceptionHandler(Exception.class)
     @ResponseStatus(HttpStatus.INTERNAL_SERVER_ERROR)
     public ResponseEntity<ErrorResponseDTO> handleAllUncaughtExceptions(Exception exception, WebRequest request) {
         log.error("Unknown error occurred", exception);
-        return ResponseEntity.internalServerError().body(new ErrorResponseDTO().error("UNEXPECTED ERROR OCCURED").exception(exception.getMessage()).status(BigDecimal.valueOf(500)));
+        return ResponseEntity.internalServerError().body(new ErrorResponseDTO().error("UNEXPECTED ERROR OCCURED").exception(exception.getMessage()).status(500));
     }
 
     @ExceptionHandler(RuntimeException.class)
@@ -51,7 +50,7 @@ public class ResponseExceptionHandler extends ResponseEntityExceptionHandler{
             return manageUserException((UserException) exception, request);
         } else {
             log.error("Uncaught RuntimeException", exception);
-            return ResponseEntity.internalServerError().body(new ErrorResponseDTO().error("UNEXPECTED ERROR OCCURED").exception(exception.getMessage()).status(BigDecimal.valueOf(500)));
+            return ResponseEntity.internalServerError().body(new ErrorResponseDTO().error("UNEXPECTED ERROR OCCURED").exception(exception.getMessage()).status(500));
         }
     }
 
@@ -59,7 +58,7 @@ public class ResponseExceptionHandler extends ResponseEntityExceptionHandler{
     @ResponseStatus(HttpStatus.BAD_REQUEST)
     public ResponseEntity<ErrorResponseDTO> handleNoSuchElementException(NoSuchElementException ex, WebRequest request) {
         return ResponseEntity.badRequest().body(new ErrorResponseDTO().error("BAD REQUEST")
-                                .exception(ex.getMessage()).status(BigDecimal.valueOf(400)));
+                                .exception(ex.getMessage()).status(400));
     }
 
     // Construct dynamic error response
@@ -70,27 +69,27 @@ public class ResponseExceptionHandler extends ResponseEntityExceptionHandler{
                 case BAD_REQUEST:
                     log.info("bad request: " + ex.getMessage());
                     return ResponseEntity.badRequest().body(new ErrorResponseDTO().error("BAD REQUEST")
-                            .exception(ex.getMessage()).status(BigDecimal.valueOf(400)));
+                            .exception(ex.getMessage()).status(400));
                 case UNAUTHORIZED:
                     log.info("user not authorized to access api: " + ex.getMessage());
                     return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(new ErrorResponseDTO().error("UNAUTHORIZED")
-                            .exception(ex.getMessage()).status(BigDecimal.valueOf(401)));
+                            .exception(ex.getMessage()).status(401));
                 case FORBIDDEN:
                     log.info("forbidden to access api: " + ex.getMessage());
                     return ResponseEntity.status(HttpStatus.FORBIDDEN).body(new ErrorResponseDTO().error("FORBIDDEN")
-                            .exception(ex.getMessage()).status(BigDecimal.valueOf(403)));
+                            .exception(ex.getMessage()).status(403));
                 case NOT_FOUND:
                     log.info("Resurce not found: " + ex.getMessage());
                     return ResponseEntity.status(HttpStatus.NOT_FOUND).body(new ErrorResponseDTO().error("NOT FOUND")
-                            .exception(ex.getMessage()).status(BigDecimal.valueOf(404)));
+                            .exception(ex.getMessage()).status(404));
                 default:
                     log.error("Uncaught HttpClientErrorException", ex);
                     return ResponseEntity.internalServerError().body(new ErrorResponseDTO().error("UNEXPECTED ERROR OCCURED")
-                            .exception(ex.getMessage()).status(BigDecimal.valueOf(500)));
+                            .exception(ex.getMessage()).status(500));
         } else {
             log.error("Uncaught UserException", ex);
             return ResponseEntity.internalServerError().body(new ErrorResponseDTO().error("UNEXPECTED ERROR OCCURED")
-                    .exception(ex.getMessage()).status(BigDecimal.valueOf(500)));
+                    .exception(ex.getMessage()).status(500));
         }
     }
 }
